+++ crates/module-system/module-implementations/sov-evm/src/authenticate.rs
@@ -128,6 +128,7 @@ where
         credential_id,
         credentials,
         default_address: S::Address::from_vm_address(ethereum_address),
+        requested_address: None,
     }
 }