+++ crates/module-system/sov-modules-api/src/transaction/mod.rs
@@ -65,6 +65,9 @@ pub struct Version0<Call, S: Spec, C: CryptoSpecExt = <S as Spec>::CryptoSpec> {
     #[serde(with = "hex_field_format")]
     #[sov_wallet(display = "hex")]
     pub pub_key: C::PublicKey,
+    /// The address of the target of the transaction. If no address is provided, the transaction is sent to the default address associated with the public key.
+    #[serde(default)]
+    pub address_override: Option<S::Address>,
     /// The runtime call of the transaction.
     #[sov_wallet(
         bound = "Call: sov_rollup_interface::sov_universal_wallet::schema::UniversalWallet"
@@ -136,6 +139,9 @@ pub struct Version1<Call, S: Spec, C: CryptoSpecExt = <S as Spec>::CryptoSpec> {
     ///  - If an invalid signature is detected, the transaction will be invalid. This allows us to use batch verification.
     // This is used to compute the credential ID statelessly.
     pub min_signers: u8,
+    /// The address of the target of the transaction. If no address is provided, the transaction is sent to the default address associated with the public key.
+    #[serde(default)]
+    pub address_override: Option<S::Address>,
     /// The runtime call of the transaction.
     #[sov_wallet(
         bound = "Call: sov_rollup_interface::sov_universal_wallet::schema::UniversalWallet"
@@ -164,6 +170,8 @@ impl<Call: BorshSerialize, S: Spec, C: CryptoSpecExt> Version1<Call, S, C> {
     /// Serializes only the `UnsignedTransaction` part of the transaction and appens the chain_hash
     pub fn serialize_for_signing(&self, chain_hash: &[u8; 32]) -> Vec<u8> {
         let mut out = Vec::with_capacity(64); // Preallocate a little capacity to avoid excessive reallocations
+        BorshSerialize::serialize(&self.address_override, &mut out)
+            .expect("Serialization to vec is infallible");
         BorshSerialize::serialize(&self.runtime_call, &mut out)
             .expect("Serialization to vec is infallible");
         BorshSerialize::serialize(&self.uniqueness, &mut out)
@@ -416,6 +424,7 @@ impl<R: TransactionCallable, S: Spec, C: CryptoSpecExt> Transaction<R, S, C> {
         details: TxDetails<S>,
     ) -> Self {
         Self::V0(Version0 {
+            address_override: None,
             signature,
             pub_key,
             runtime_call,
@@ -469,15 +478,17 @@ impl<R: TransactionCallable, S: Spec, C: CryptoSpecExt> Transaction<R, S, C> {
     /// Converts the transaction to an unsigned transaction.
     pub fn to_unsigned_transaction(&self) -> UnsignedTransaction<R, S> {
         match &self {
-            Transaction::V0(inner) => UnsignedTransaction::new_with_details(
+            Transaction::V0(inner) => UnsignedTransaction::new_with_details_and_address_override(
                 inner.runtime_call.clone(),
                 inner.uniqueness,
                 inner.details.clone(),
+                inner.address_override.clone(),
             ),
-            Transaction::V1(inner) => UnsignedTransaction::new_with_details(
+            Transaction::V1(inner) => UnsignedTransaction::new_with_details_and_address_override(
                 inner.runtime_call.clone(),
                 inner.uniqueness,
                 inner.details.clone(),
+                inner.address_override.clone(),
             ),
         }
     }
@@ -489,6 +500,9 @@ impl<R: TransactionCallable, S: Spec, C: CryptoSpecExt> Transaction<R, S, C> {
 )]
 #[serde(bound = "R::Call: serde::Serialize + serde::de::DeserializeOwned")]
 pub struct UnsignedTransaction<R: TransactionCallable, S: Spec> {
+    /// The address of the target of the transaction. If no address is provided, the transaction is sent to the default address associated with the public key.
+    #[serde(default)]
+    pub address_override: Option<S::Address>,
     /// The runtime call
     pub runtime_call: R::Call,
     /// The uniqueness identifier
@@ -530,6 +544,30 @@ impl<R: TransactionCallable, S: Spec> UnsignedTransaction<R, S> {
         gas_limit: Option<S::Gas>,
     ) -> Self {
         Self {
+            address_override: None,
+            runtime_call,
+            uniqueness,
+            details: TxDetails {
+                max_priority_fee_bips,
+                max_fee,
+                gas_limit,
+                chain_id,
+            },
+        }
+    }
+
+    /// Creates a new [`UnsignedTransaction`] with the given arguments.
+    pub const fn new_with_address_override(
+        runtime_call: R::Call,
+        chain_id: u64,
+        max_priority_fee_bips: PriorityFeeBips,
+        max_fee: Amount,
+        uniqueness: UniquenessData,
+        gas_limit: Option<S::Gas>,
+        address_override: Option<S::Address>,
+    ) -> Self {
+        Self {
+            address_override,
             runtime_call,
             uniqueness,
             details: TxDetails {
@@ -542,16 +580,27 @@ impl<R: TransactionCallable, S: Spec> UnsignedTransaction<R, S> {
     }
 
     /// Creates a new unsigned transaction with the provided metadata.
+    pub const fn new_with_details_and_address_override(
+        runtime_call: R::Call,
+        uniqueness: UniquenessData,
+        details: TxDetails<S>,
+        address_override: Option<S::Address>,
+    ) -> Self {
+        Self {
+            address_override,
+            runtime_call,
+            uniqueness,
+            details,
+        }
+    }
+
+    /// Creates a new unsigned transaction with the provided metadata.
     pub const fn new_with_details(
         runtime_call: R::Call,
         uniqueness: UniquenessData,
         details: TxDetails<S>,
     ) -> Self {
-        Self {
-            runtime_call,
-            uniqueness,
-            details,
-        }
+        Self::new_with_details_and_address_override(runtime_call, uniqueness, details, None)
     }
 
     /// Creates a new [`Transaction`] from this [`UnsignedTransaction`] when given a signature
@@ -576,6 +625,7 @@ impl<R: TransactionCallable, S: Spec> UnsignedTransaction<R, S> {
         multisig: Multisig<<S::CryptoSpec as CryptoSpec>::PublicKey>,
     ) -> Version1<R::Call, S> {
         Version1 {
+            address_override: self.address_override,
             signatures: SafeVec::new(),
             unused_pub_keys: multisig
                 .signers